service/requirements.txt
Flask
pyicloud
python-dateutil
